Make instruction counter in simulators 64-bit (issue 22302). This requires supporting uint64_t flag values.
R=zra@google.com

Review URL: https://codereview.chromium.org//898093003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@43660 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/simulator_arm.h
@@ -73,7 +73,7 @@ class Simulator {
   uword StackTop() const;
 
   // Accessor to the instruction counter.
-  intptr_t get_icount() const { return icount_; }
+  uint64_t get_icount() const { return icount_; }
 
   // The isolate's top_exit_frame_info refers to a Dart frame in the simulator
   // stack. The simulator's top_exit_frame_info refers to a C++ frame in the
@@ -156,7 +156,7 @@ class Simulator {
   // Simulator support.
   char* stack_;
   bool pc_modified_;
-  intptr_t icount_;
+  uint64_t icount_;
   static int32_t flag_stop_sim_at_;
   SimulatorSetjmpBuffer* last_setjmp_buffer_;
   uword top_exit_frame_info_;
